Publish clusters in a single pointcloud
Removed all the publishers for each cluster, and replaced with a single one that send all the clusters at once, adding a label to each point to distinguish between them.

Fixed warning related to Eigen in CMakeLists.txt

Fixed error when creating a KdTree in bounding_boxes_node with an empty cloud.

Added absolute time logging to boxes_tracker_node and added a flag to know when the first_box is received, to make the graphs start exactly at 0.