Post detection and matching improvements
/// bounding_boxes_node ///
Modified the code that looks for posts in bounding_boxes_node. Now, it has a triple nested for loop to check distances between all possible posts (boxes), and find a triangle that match the expected distances (with some margin of error). 
///////////////////////////////////
/// matching_node ///
Initial received post labeling is also improved.
Deleted some subscriptions and its associated callbacks, that are no longer needed because the post positions are received now in a single message.
Added preliminary code to generate virtual posts when only two of them are detected. We still need to debug virtual post generation.
//////////////////////////