Change to path to and between posts publishing
Before this change, the paths are only published when at least one bounding box match the dimensions of a post. When only one or two posts are detected, the rest of the paths are empty, which makes rviz delete the old paths. But when no post is detected, all the paths are empty, but the publishing was made inside an if statement, so rviz didn't refreshed the paths when no posts were detected.
Now, path publishing is made outside that if statement, so is always done, and the paths shown in rviz are refreshed.