Added rate to sensor_filter_node and bounding_boxes_node
It's not needed to process every single pointcloud received. With a rate.sleep and a message queue of depth 1, some of them are discarded. This change has an positive impact in performance, but the results are similar if a sufficiently high working frequency is configured. 10 Hz appears to yield good results, and smooth visualization in Rviz, saving computational power.

The changes done are the following:
Added rate.sleep() in main loop of sensor_filter_node and bounding_boxes_node.
Added working_frequency parameter to those nodes.
Added that parameter to launch files.