Publish all radar candidates in a single message.
Publish all radar candidates received on a certain time interval in a single message containing an array of individual messages.

Added a parameter to configure that interval, and adapted the code to accumulate radar candidates during it.

Added new messages needed, including AisCandidatesList.msg for use when needed.

Some aesthetic changes of the code.