Modified getMahaDistance method to return the square root of the result
If the square root is not calculated, the result is a squared distance, whereas the thresholds to decide if a match between a target and a candidate is valid are simple distances (not squared). So a comparison between them didn't have sense, and caused a bad behavior of the obstacle estimator.

Also changed the estimator_rate to match the one at which the rest of the nodes work, but especially, the one at which bounding_boxes_node publish messages at the topic lidar_candidates.