Some changes related with the use of simulated time
Added a while(ros::Time::now() == 0) to the nodes to wait for reception of first message through /clock topic when in simulation mode.

Changed target_tracker timer to use ros::Time instead of std::chrono, so the estimator can be used in simulations, and those ones can now be paused, and the timer will not keep increasing.